Hoàn tất phân quyền cho Mod

+++ trunk/Elibrary/src/java/uit/elib/action/DeleteSubjectAction.java
@@ -8,12 +8,14 @@ import java.io.File;
 import java.util.List;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
 import org.apache.struts.action.ActionForm;
 import org.apache.struts.action.ActionForward;
 import org.apache.struts.action.ActionMapping;
 import uit.elib.bo.ResourceBO;
 import uit.elib.bo.SubjectBO;
 import uit.elib.dto.Resource;
+import uit.elib.utility.CheckGroup;
 /**
  *
  * @author HERO
@@ -35,27 +37,41 @@ public class DeleteSubjectAction extends org.apache.struts.action.Action {
     public ActionForward execute(ActionMapping mapping, ActionForm form,
             HttpServletRequest request, HttpServletResponse response)
             throws Exception {
+        int checkgroup =2; //visitor 
+        HttpSession session = request.getSession();
+        if(session.getAttribute("username")!=null){ 
+            CheckGroup checkGroup = new CheckGroup();
+            checkgroup = checkGroup.Group((String)session.getAttribute("username"));
+            if(checkgroup==-1) // account has just been locked while users are accessing or  account has just expired while users are accessing
+            {   
+                 session.removeAttribute("username");
+                 session.removeAttribute("group");
+            }
+            if(checkgroup==1||checkgroup==3)// admin or mod
+            {               
                 String subjectsID = request.getParameter("subjectsID");
-        subjectsID = subjectsID.substring(0, subjectsID.length()-1);
-        String sqlresource = "delete from resource where SubjectID in("+subjectsID+")";
-        String sqlsubject = "delete from subject where SubjectID in("+subjectsID+")";
-        List<Resource> listresource = ResourceBO.getResourceBO().getAllResource("SubjectID in("+subjectsID+")", null);
-//        ResourceBO resourceBO = ResourceBO.getResourceBO();
-//        resourceBO.DeleteResource(sqlresource);
-//        SubjectBO subjectBO = SubjectBO.getSubjectBO();
-//        subjectBO.DeleteSubject(sqlsubject);
-        for (int i = 0; i < listresource.size(); i++) 
-        {
-            if(listresource.get(i).getServerName()!=null)
-            {
-                String filePath = getServlet().getServletContext().getRealPath("/") +"upload/"+listresource.get(i).getServerName();
-                File file = new File(filePath);
-                file.delete();
+                subjectsID = subjectsID.substring(0, subjectsID.length()-1);
+                String sqlresource = "delete from resource where SubjectID in("+subjectsID+")";
+                String sqlsubject = "delete from subject where SubjectID in("+subjectsID+")";
+                List<Resource> listresource = ResourceBO.getResourceBO().getAllResource("SubjectID in("+subjectsID+")", null);
+        //        ResourceBO resourceBO = ResourceBO.getResourceBO();
+        //        resourceBO.DeleteResource(sqlresource);
+        //        SubjectBO subjectBO = SubjectBO.getSubjectBO();
+        //        subjectBO.DeleteSubject(sqlsubject);
+                for (int i = 0; i < listresource.size(); i++) 
+                {
+                    if(listresource.get(i).getServerName()!=null)
+                    {
+                        String filePath = getServlet().getServletContext().getRealPath("/") +"upload/"+listresource.get(i).getServerName();
+                        File file = new File(filePath);
+                        file.delete();
+                    }
+                }
+
+                ResourceBO.getResourceBO().DeleteResource(sqlresource);
+                SubjectBO.getSubjectBO().DeleteSubject(sqlsubject);
             }
         }
-        
-        ResourceBO.getResourceBO().DeleteResource(sqlresource);
-        SubjectBO.getSubjectBO().DeleteSubject(sqlsubject);
-        return mapping.findForward(null);
+        return null;
     }
 }
